requirements.txt
anyio==4.11.0
certifi==2025.11.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-dotenv==1.2.1
python-telegram-bot==21.6
sniffio==1.3.1
typing_extensions==4.15.0
